requirements.txt
ipython==8.6.0
netCDF4==1.6.1
pandas==1.5.1
rasterio==1.3.3
xarray==2022.10.0
scipy
tqdm
pyproj
matplotlib==3.7.5
tensorflow_addons==0.21
scikit-learn==1.3.2
hvplot
bokeh
